Upgrade to Heroicons v2
This changes the `Solid` icon to have a viewBox of `0 0 24 24` (instead of the old `0 0 20 20`) and introduces a new `Mini` variant which is like the old `Solid`.

Also, basically all icons are either changed, removed, or newly added. So this is gomponents-heroicons v2.

See the fork for building this at https://github.com/maragudk/heroicons